refactor(engine): split concurrent_delta parallel_apply into submodules
